--- packages/core/echo/echo-schema/src/handler/index.ts
@@ -1,7 +1,0 @@
-//
-// Copyright 2024 DXOS.org
-//
-
-export * from './expando';
-export * from './ref';
-export * from './utils';
