+++ packages/core/echo/echo-schema/src/object/index.ts
@@ -4,3 +4,4 @@
 
 export * from './typed-object';
 export * from './object-id';
+export * from './expando';
